requirements.txt
absl-py==2.1.0
affine==2.4.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
augmentations==0.0.1
Babel==2.13.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.2
bleach==6.1.0
blosc2==2.5.1
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
brotlipy==0.7.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
cfgv==3.4.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.86
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
colorlog==6.8.2
comm==0.2.0
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cryptography @ file:///C:/b/abs_36x9ifdcl4/croot/cryptography_1665612655344/work
cupy-cuda116==10.6.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.0.11
DCLS==0.1.1
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
distlib==0.3.8
docker-pycreds==0.4.0
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
etils==1.5.2
exceptiongroup==1.2.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema==2.19.0
fastrlock==0.8.2
filelock==3.14.0
flatbuffers==24.3.25
flax==0.8.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.25.0
fqdn==1.5.1
fsspec==2024.5.0
gast==0.6.0
geojson==3.1.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
huggingface-hub==0.23.1
humanize==4.10.0
identify==2.6.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
imageio==2.34.2
importlib-metadata==7.0.0
importlib_resources==6.4.4
intel-openmp==2021.4.0
ipykernel @ file:///C:/b/abs_21ykzkm7y_/croots/recipe/ipykernel_1662361803478/work
ipython @ file:///C:/b/abs_536xsn7rj0/croot/ipython_1670919360443/work
ipywidgets==8.1.1
isoduration==20.11.0
jax==0.4.30
jaxlib==0.4.30
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client @ file:///C:/b/abs_cbnezz4zg0/croot/jupyter_client_1669040266195/work
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.4.1
kiwisolver @ file:///C:/ci/kiwisolver_1653292407425/work
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib @ file:///C:/b/abs_03z_dj2gty/croot/matplotlib-suite_1670466145509/work
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661915841596/work
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
munkres==1.1.4
namex==0.0.8
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
ndindex==1.8
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929390/work
networkx==3.2.1
nodeenv==1.9.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.3
optree==0.11.0
orbax-checkpoint==0.6.1
overrides==7.4.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/b/abs_c6fuezktfm/croot/pandas_1670425103552/work
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
platformdirs==4.1.0
ply==3.11
pre-commit==3.7.1
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.18.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PyQt5==5.15.7
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
pysnn @ file:///D:/pytorch/reproduction/FedSNN-master/pysnn-0.1.1/pysnn-0.1.1/pysnn-0.1.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
PySoundFile==0.9.0.post1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
pywin32==305.1
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rasterio==1.3.10
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
sentry-sdk==2.7.1
setproctitle==1.3.3
shapely==2.0.5
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
spikingjelly==0.0.0.0.14
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.12
tables==3.9.2
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.64
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.7.24
timm==0.4.12
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
toolz==0.12.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado @ file:///C:/ci/tornado_1662458743919/work
tqdm==4.66.1
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
turfpy==0.0.7
types-python-dateutil==2.8.19.14
typing_extensions==4.12.0
uri-template==1.3.0
urllib3 @ file:///C:/b/abs_21qa0j12xt/croot/urllib3_1670527000672/work
virtualenv==20.26.3
wandb==0.17.3
wandb_osh==1.2.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.9
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt==1.16.0
zipp==3.17.0
